+++ CHANGES.md
@@ -5,7 +5,7 @@ All notable changes to Prescrypt will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.1] - unreleased
+## [0.9.1] - 2026-01-19
 
 ### Added
 
@@ -73,7 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Statistics
 
-- 2403 tests passing
+- 2375 tests passing
+- 89.33% test coverage (cf. `make test-cov`)
 
 
 ## [0.9.0] - 2026-01-16